Weights are now floats; Other small fixes.
--HG--
rename : src/main/java/org/geolatte/graph/algorithms/EmptyContextualReachability.java => src/main/java/org/geolatte/graph/EmptyContextualReachability.java

+++ src/main/java/org/geolatte/graph/Locatable.java
@@ -32,10 +32,10 @@ public interface Locatable {
     /**
      * @return The X-coordinate
      */
-    public int getX();
+    public float getX();
 
     /**
      * @return The Y-coordinate
      */
-    public int getY();
+    public float getY();
 }
